chore: MenuBarの定義を分散
